update to playwright tests
