--- .github/actions/ete-docs-bundle/action.yml
@@ -1,45 +1,0 @@
-name: E2E @fern-ui/docs-bundle
-description: Run Playwright tests on the @fern-ui/docs-bundle package
-
-inputs:
-  deployment_url:
-    description: "The URL of the deployment to test"
-    required: true
-  token:
-    description: "The Vercel token to use for the deployment"
-    required: true
-  fern_token:
-    description: "The token to use for the Fern API"
-    required: true
-
-runs:
-  using: "composite"
-  steps:
-    - uses: ./.github/actions/install
-
-    - shell: bash
-      name: Fetch domains
-      run: pnpm vercel-scripts domains.txt ${{ inputs.deployment_url }} --token=${{ inputs.token }}
-
-    - name: Install Playwright Browsers
-      shell: bash
-      run: pnpm exec playwright install --with-deps
-
-    - name: Run Playwright tests
-      shell: bash
-      env:
-        DEPLOYMENT_URL: ${{ inputs.deployment_url }}
-        PLAYWRIGHT_JSON_OUTPUT_NAME: results.json
-      run: pnpm exec playwright test playwright/smoke --workers 6 --reporter json
-
-    - name: Run E2E tests
-      shell: bash
-      env:
-        DEPLOYMENT_URL: ${{ inputs.deployment_url }}
-        FERN_TOKEN: ${{ inputs.fern_token }}
-      run: pnpm i -g fern-api@latest; pnpm docs:e2e
-
-    - uses: daun/playwright-report-summary@v3
-      if: always()
-      with:
-        report-file: results.json
