+++ playwright/subpath-proxy/index.spec.ts-snapshots/sitemap-xml-should-match-snapshot-1-chromium-darwin.txt
@@ -1,7 +1,4 @@
-// Vitest Snapshot v1, https://vitest.dev/guide/snapshot.html
-
-exports[`nginx-proxy > sitemap.xml should match snapshot 1`] = `
-"<?xml version="1.0" encoding="UTF-8"?> <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+<?xml version="1.0" encoding="UTF-8"?> <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
         
                     <url>
                         <loc>https://test-nginx-proxy.docs.buildwithfern.com/subpath/test-1</loc>
@@ -12,5 +9,4 @@ exports[`nginx-proxy > sitemap.xml should match snapshot 1`] = `
                         <loc>https://test-nginx-proxy.docs.buildwithfern.com/subpath/capture-the-flag</loc>
                     </url>
                 
-    </urlset>"
-`;
+    </urlset>
